# filmApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API",
        "description": "API de gestion des films, acteurs, critiques et utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
        "/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Supprimer une critique",
                "operationId": "791256020eb951289869f5eef122da5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la critique",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Critique supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Critique supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Critique introuvable"
                    }
                }
            }
        },
...
    },
...
}